fix: process extraction task not shifting pointer correctly
When we get a chunk of resource records at each scheduled task to
create extractions for, the starting resource id should shift
by the maximum process count so the next scheduled task for this
resource knows where to start.
Fixed an issue where the id was always being reset to 0 (zero) and
starting from the beginning of table again.